E2E: Move database roundtrip tests to their own file
